Removed duplicate repositories
