--- GirafRest/IRepositories/IActivityRepository.cs
@@ -1,12 +1,0 @@
-﻿using GirafRest.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq.Expressions;
-using System.Threading.Tasks;
-
-namespace GirafRest.IRepositories
-{
-    public interface IActivityRepository : IRepository<Activity>
-    {
-    }
-}
