--- GirafRest/IRepositories/IPictogramRelationRepository.cs
@@ -1,13 +1,0 @@
-using GirafRest.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq.Expressions;
-using System.Threading.Tasks;
-
-namespace GirafRest.IRepositories
-{
-    public interface IPictogramRelationRepository : IRepository<PictogramRelation>
-    {
-        public ICollection<PictogramRelation> GetWithPictogram(int activityID);
-    }
-}
