+++ GirafRest/IRepositories/IRepositories M to M/IActivityRepository.cs
@@ -1,11 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq.Expressions;
 using GirafRest.Models;
 namespace GirafRest.IRepositories
 {
-    public interface IActivityRepository : IRepository<Models.Activity>
+    public interface IActivityRepository : IRepository<Activity>
     {
 
     }
