--- GirafRest/Repositories/ActivityRepository.cs
@@ -1,17 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using Microsoft.EntityFrameworkCore;
-using GirafRest.Models;
-using GirafRest.IRepositories;
-using GirafRest.Data;
-using System.Threading.Tasks;
-
-namespace GirafRest.Repositories
-{
-    public class ActivityRepository : Repository<Activity>, IActivityRepository
-    {
-        public ActivityRepository(GirafDbContext context) : base(context)
-        {
-        }
-    }
-}
